Improve name validation
The parser finds all known prefixes in an input string even when a
prefix is inside the value of another field. This prevents users from
entering names with text resembling a prefix and creates a poor user
experience. Moreover, only alphanumerical characters are allowed which
is overly restrictive for certain names that contain hyphens or
apostrophes.

Let's:
* expand the regex to allow Unicode and a set of special
non-alphanumerical characters that are commonly found in names
* allow users to enclose names with prefixes with one set of double
quotes
* capitalize letters following any of these new special characters for
better formatting

Making the tokenizer quote-aware is the most robust solution to this
problem, because it allows the Name field to be fully flexible without
requiring users to avoid certain characters.

A small side effect that results from this consistent naming standard
is that names with multiple capitalised letters will only have their
first letter capitalised.